Update version and hash types

+++ csrc/object.c
@@ -74,7 +74,7 @@ ObjNative* newNative(NativeFn function) {
     return native;
 }
 
-ObjString* allocateString(char* chars, int length, uint32_t hash){
+ObjString* allocateString(char* chars, int length, uint64_t hash){
     ObjString* string = ALLOCATE_OBJ(ObjString, OBJ_STRING);
     string->length = length;
     string->chars = chars;
@@ -85,8 +85,8 @@ ObjString* allocateString(char* chars, int length, uint32_t hash){
     return string;
 }
 
-uint32_t hashString(const char* key, int length){
-    uint32_t hash = 2166136261u;
+uint64_t hashString(const char* key, int length){
+    uint64_t hash = 2166136261u;
     for(int i = 0; i < length; i++){
         hash ^= (uint8_t)key[i];
         hash *= 16777619;
@@ -96,7 +96,7 @@ uint32_t hashString(const char* key, int length){
 }
 
 ObjString* takeString(char* chars, int length){
-    uint32_t hash = hashString(chars, length);
+    uint64_t hash = hashString(chars, length);
     ObjString* interned = tableFindString(&vm.strings, chars, length, hash);
     if(interned != NULL){
         FREE_ARRAY(char, chars, length+1);
@@ -106,7 +106,7 @@ ObjString* takeString(char* chars, int length){
 }
 
 ObjString* copyString(const char* chars, int length){
-    uint32_t hash = hashString(chars, length);
+    uint64_t hash = hashString(chars, length);
     ObjString* interned = tableFindString(&vm.strings, chars, length, hash);
     if(interned != NULL) return interned;
     char* heapChars = ALLOCATE(char, length + 1);
